Use Selleckfile.js to register Handlebars helpers.
Find the first file named 'Selleckfile.js' in the current directory or
the nearest ancestor directory and invoke its exported function with our
Handlebars implementation to allow for custom helpers to be registered.

The filename 'Selleckfile.js' is not case-sensitive.

The interface exposed to the Selleckfile function is an object with the
single property `handlebars`. The Selleckfile is expected to invoke
`handlebars.registerHelper` as required to add custom helpers.